+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./model.js";
 export * from "./boundary.js";
 export * from "./normalize.js";
+export * from "./semantic.js";
 export * from "./graph.js";